+++ sphinxcontrib/bibtex/__init__.py
@@ -26,10 +26,6 @@ def init_bibtex_cache(app):
     """
     if not hasattr(app.env, "bibtex_cache"):
         app.env.bibtex_cache = Cache()
-    # XXX labels are currently not cached, always calculated on the fly
-    app.env.bibtex_citation_label = {}
-    # XXX same for list of cited references
-    app.env.bibtex_cited = set()
 
 def purge_bibtex_cache(app, env, docname):
     """Remove all information related to *docname* from the cache.
@@ -42,7 +38,7 @@ def purge_bibtex_cache(app, env, docname):
     env.bibtex_cache.purge(docname)
 
 def process_citations(app, doctree, docname):
-    """Replace labels of citation nodes by numbers.
+    """Replace labels of citation nodes by actual labels.
 
     :param app: The sphinx application.
     :type app: :class:`sphinx.application.Sphinx`
@@ -52,16 +48,16 @@ def process_citations(app, doctree, docname):
     :type docname: :class:`str`
     """
     for node in doctree.traverse(docutils.nodes.citation):
-        label = node[0].astext()
+        key = node[0].astext()
         try:
-            num = app.env.bibtex_citation_label[label]
+            label = app.env.bibtex_cache.get_label_from_key(key)
         except KeyError:
-            num = str(len(app.env.bibtex_citation_label) + 1)
-            app.env.bibtex_citation_label[label] = num
-        node[0] = docutils.nodes.label('', num)
+            app.warn("could not relabel citation [%s]" % key)
+        else:
+            node[0] = docutils.nodes.label('', label)
 
 def process_citation_references(app, doctree, docname):
-    """Replace text of citation reference nodes by numbers.
+    """Replace text of citation reference nodes by actual labels.
 
     :param app: The sphinx application.
     :type app: :class:`sphinx.application.Sphinx`
@@ -75,9 +71,13 @@ def process_citation_references(app, doctree, docname):
     for node in doctree.traverse(docutils.nodes.reference):
         text = node[0].astext()
         if text.startswith('[') and text.endswith(']'):
-            label = text[1:-1]
-            node[0] = docutils.nodes.Text(
-                '[' + app.env.bibtex_citation_label[label] + ']')
+            key = text[1:-1]
+            try:
+                label = app.env.bibtex_cache.get_label_from_key(key)
+            except KeyError:
+                app.warn("could not relabel citation reference [%s]" % key)
+            else:
+                node[0] = docutils.nodes.Text('[' + label + ']')
 
 def setup(app):
     """Set up the bibtex extension: