+++ test/test_issue1.py
@@ -19,7 +19,6 @@ def teardown_module():
 @with_app(srcdir=srcdir, warningiserror=True)
 def test_tinker(app):
     app.builder.build_all()
+    nose.tools.assert_true(app.env.bibtex_cache.is_cited(u"2011:BabikerIPv6"))
     nose.tools.assert_equal(
-        app.env.bibtex_cited, set([u"2011:BabikerIPv6"]))
-    nose.tools.assert_equal(
-        app.env.bibtex_citation_label, {u"2011:BabikerIPv6": "1"})
+        app.env.bibtex_cache.get_label_from_key(u"2011:BabikerIPv6"), u"1")